+++ packages/cli/src/commands/generate.ts
@@ -73,11 +73,7 @@ export async function generate(opts: GenerateOptions) {
           await Promise.all(
             Object.entries(config[kind] ?? {}).map(async ([key, v]) => [
               key,
-              toHex(
-                Blake2128(
-                  Binary.fromText(await fs.readFile(v, "utf-8")).asBytes(),
-                ),
-              ),
+              toHex(Blake2128(Binary.fromText(await fs.readFile(v, "utf-8")))),
             ]),
           ),
         ),
@@ -302,9 +298,7 @@ async function outputCodegen(
   )
 
   const hash = h64(
-    Binary.fromText(
-      Array.from(metadataTypes.checksumToIdx.keys()).join(""),
-    ).asBytes(),
+    Binary.fromText(Array.from(metadataTypes.checksumToIdx.keys()).join("")),
   )
 
   const EntryPointsCodec = Vector(EntryPointCodec)