+++ packages/client/CHANGELOG.md
@@ -23,11 +23,16 @@
 - BREAKING: When creating transactions the optional `at` value only accepts specific block-hashes.
 - BREAKING: `getBody` and `getHeader` only accept block-hashes.
 - BREAKING: renamed `watchBlockBody` to `getBlockBody$`
+- BREAKING: `getBody` and `getBlockBody$` return `Uint8Array`s instead of `HexString`
+- BREAKING: `sign` and `getBareTx` return `Uint8Array`s instead of `HexString`
+- BREAKING: `watchValue` options are now consistent with `watchEntries` options, and also with the rest of query options.
 - BREAKING: `watchValue` emits an object that contains the value and the block the value was found in.
 - BREAKING: `watchValue` emits every time it queries the value. The value keeps the same reference if it didn't change.
 - BREAKING: rename chainspecs `ksmcc3` => `kusama`, `westend2` => `westend`
 - BREAKING: flattened Event interface in TypedAPI event `watch`.
 - BREAKING: `filter()` method in TypedAPI event entry returns an object `{ original: SystemEvent, payload: T }` instead of just the payload.
+- BREAKING: Codegen exposes and expects `Vec<u8>` as `Uint8Array` instead of `Binary`.
+- BREAKING: Remove `Binary` class. Replaced by a set of utilities to deal with `Uint8Array`s.
 
 - **WS-Provider:**
   - BREAKING: This export has been moved in favor of `polkadot-api/ws`