+++ packages/observable-client/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changed
 
+- `body$` and `call$` return `Uint8Array` instead of HexString.
+- `broadcastTx$` takes a `Uint8Array` as a parameter instead of HexString.
 - `storage$` returns `Observable<{ block: BlockInfo, value: T }>` instead of `Observable<T>`
 
 ## Unreleased